src/requirements.txt
accelerate==0.33.0
annotated-types==0.7.0
APScheduler==3.10.4
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2024.7.4
charset-normalizer==3.3.2
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.23.5
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.8.0.76
packaging==24.1
pillow==10.0.1
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
safetensors==0.4.3
six==1.16.0
sympy==1.13.0
timm==0.9.16
tokenizers==0.19.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
